Added the file .fvwm2rc_DARK
This is my "dark theme" developed in April 2020, and while working only from
my home office, which is pretty dark.

This dark mode uses purple window frames, and light gray for highlighting. It has
four XTerm executions in the popup menu, two of which include the retro style green
and orange terminal colours. The xload, desktop pager, and window buttons in the
Buttons box use a purple background and mostly gray foreground colours. The xlock
that is in the Buttons panel uses a dark gray background and black foreground dials.
IN 2020/04/21 "master"